MIT 模式实测通过 + 串口 mit0~7 测试序列
- MIT 速度反馈拆 LPF 状态到 controller (mit_vel_filt/mit_active), 切入复位避免残值阶跃
- MIT 模式跳过电流环斜坡 (CurrentLoopSmoothRun), 直通 iq_basic 保证响应
- dtheta_mech_out 修正为输出端 rpm*1024 (去掉多乘的减速比 25)
- 0x500 MIT 帧解析加临界区保护 + MIT_TIMEOUT_MS=20 看门狗
- can_wly 超时保护默认启用 (g_can_timeout_force_disable=1 调试态)
- foc_bsp.c 新增 mit0~7 串口命令: Step0 安全态 + Step1~7 MIT 序列 (10 次 @10ms)
- CAN_WLY_PROTOCOL_VERIFY.md 加 MIT 实测验证序列章节

+++ Core/Src/main.c
@@ -150,7 +150,7 @@ int main(void)
 	printf("Calibrating ADC offsets...\r\n");
 	HAL_Delay(100);  // 等待ADC稳定
 	ADC_CalibrateOffsets(1024);
-	printf("ADC calibration done: Off_a=%ld Off_b=%ld\r\n",
+	printf("ADC calibration done: Off_a=%d Off_b=%d\r\n",
 	       (int32_t)g_adc_offset_a, (int32_t)g_adc_offset_b);
 
 	/* FOC控制器初始化（参考PHU Init_func顺序） */
@@ -186,8 +186,8 @@ int main(void)
 	HAL_Delay(10);
 	/* VDC: ADC 16bit, 3.3V 满量程, 分压比需根据硬件确认 */
 	float vdc_volt = g_vdc_raw * 3.3f / 65535.0f;
-	printf("VDC ADC raw = %lu, voltage (before divider scaling) = %.3f V\r\n",
-	       g_vdc_raw, vdc_volt);
+	printf("VDC ADC raw = %u, voltage (before divider scaling) = %.3f V\r\n",
+	       (unsigned int)g_vdc_raw, vdc_volt);
 
 //	/* Rs 辨识前先对齐 d 轴，消除上电编码器位置不确定 */
 //	alignDAxis(&controller_eyou);
